Show PoC for obtaining CPU time used by process via shared memory

+++ server.cpp
@@ -24,10 +24,6 @@
 
 #include "memfd.hpp"
 
-static void error(const char *msg) {
-  perror(msg);
-  exit(EXIT_FAILURE);
-}
 
 static int new_memfd_region(char *unique_str) {
   char *shm;
@@ -35,34 +31,28 @@ static int new_memfd_region(char *unique_str) {
   int fd, ret;
 
   fd = memfd_create("Server memfd", MFD_ALLOW_SEALING);
-  if (fd == -1)
-    error("memfd_create()");
+  if (fd == -1) error("memfd_create()");
 
   ret = ftruncate(fd, shm_size);
-  if (ret == -1)
-    error("ftruncate()");
+  if (ret == -1) error("ftruncate()");
 
   ret = fcntl(fd, F_ADD_SEALS, F_SEAL_SHRINK);
-  if (ret == -1)
-    error("fcntl(F_SEAL_SHRINK)");
+  if (ret == -1) error("fcntl(F_SEAL_SHRINK)");
 
   shm = static_cast<char *>(mmap(NULL, shm_size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0));
-  if (shm == MAP_FAILED)
-    error("mmap()");
+  if (shm == MAP_FAILED) error("mmap()");
 
   sprintf(shm, "Secure zero-copy message from server: %s", unique_str);
 
   /* Seal writes too, but unmap our shared mappings beforehand */
   ret = munmap(shm, shm_size);
-  if (ret == -1)
-    error("munmap()");
-  ret = fcntl(fd, F_ADD_SEALS, F_SEAL_WRITE);
-  if (ret == -1)
-    error("fcntl(F_SEAL_WRITE)");
+  if (ret == -1) error("munmap()");
+//  ret = fcntl(fd, F_ADD_SEALS, F_SEAL_WRITE);
+//  if (ret == -1)
+//    error("fcntl(F_SEAL_WRITE)");
 
   ret = fcntl(fd, F_ADD_SEALS, F_SEAL_SEAL);
-  if (ret == -1)
-    error("fcntl(F_SEAL_SEAL)");
+  if (ret == -1) error("fcntl(F_SEAL_SEAL)");
 
   return fd;
 }
@@ -103,8 +93,7 @@ static void send_fd(int conn, int fd) {
   *((int *) CMSG_DATA(CMSG_FIRSTHDR(&msgh))) = fd;
 
   int size = sendmsg(conn, &msgh, 0);
-  if (size < 0)
-    error("sendmsg()");
+  if (size < 0) error("sendmsg()");
 }
 
 #define LOCAL_SOCKET_NAME    "./unix_socket"
@@ -117,43 +106,49 @@ static void start_server_and_send_memfd_to_clients() {
 
 
   sock = socket(PF_UNIX, SOCK_STREAM, 0);
-  if (sock == -1)
-    error("socket()");
+  if (sock == -1) error("socket()");
 
   memset(&address, 0, sizeof(address));
   address.sun_family = AF_UNIX;
   snprintf(address.sun_path, UNIX_PATH_MAX, LOCAL_SOCKET_NAME);
 
   ret = unlink(LOCAL_SOCKET_NAME);
-  if (ret != 0 && ret != -ENOENT && ret != -EPERM)
-    error("unlink()");
+  if (ret != 0 && ret != -ENOENT && ret != -EPERM) error("unlink()");
 
   ret = bind(sock, (struct sockaddr *) &address, sizeof(address));
-  if (ret != 0)
-    error("bind()");
+  if (ret != 0) error("bind()");
 
   ret = listen(sock, MAX_CONNECT_BACKLOG);
-  if (ret != 0)
-    error("listen()");
+  if (ret != 0) error("listen()");
 
-  while (true) {
-    conn = accept(sock, (struct sockaddr *) &address, &addrlen);
-    if (conn == -1)
-      break;
+//  while (true) {
+  conn = accept(sock, (struct sockaddr *) &address, &addrlen);
+//    if (conn == -1)
+//      break;
 
-    /* Remove useless ctime(3) trailing newline */
-    time_t now = time(NULL);
-    char *nowbuf = ctime(&now);
-    nowbuf[strlen(nowbuf) - 1] = '\0';
+  /* Remove useless ctime(3) trailing newline */
+  time_t now = time(nullptr);
+  char *nowbuf = ctime(&now);
+  nowbuf[strlen(nowbuf) - 1] = '\0';
 
-    printf("[%s] New client connection!\n", nowbuf);
+  printf("[%s] New client connection!\n", nowbuf);
 
-    fd = new_memfd_region(nowbuf);
-    send_fd(conn, fd);
+  fd = new_memfd_region(nowbuf);
+  send_fd(conn, fd);
 
-    close(conn);
-    close(fd);
+  char *shm;
+  const int shm_size = 1024;
+
+  shm = static_cast<char *>(mmap(nullptr, shm_size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0));
+  if (shm == MAP_FAILED) errorp("mmap");
+
+  while (1) {
+    sleep(5);
+    printf("Received from client: %s\n", shm);
   }
+
+  close(conn);
+  close(fd);
 }
 
 int main(int argc, char **argv) {
